Manejo de errores
majeno de errores basicamente

+++ php forms/delete/DCentrosVerificacion.php
@@ -1,17 +1,32 @@
 <?php
-    //Obtener datos
-    $IdPropietario = $_POST['ID'];
+        if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+            $IdPropietario = $_POST['ID'];
 
-    $SQL= "DELETE FROM CENTROSVERIFICACION WHERE ID = '$IdPropietario'";
+            include("../../controlador/controlador.php");
 
-    include("../../controlador/controlador.php");
-    $Con = Conectar();
-    $ResultSet = Ejecutar($Con,$SQL);
-    Desconectar($Con);
+            try {
+                $Con = Conectar();
+                $checkSQL = "SELECT * FROM CENTROSVERIFICACION WHERE ID = '$IdPropietario'";
+                $checkResult = Ejecutar($Con, $checkSQL);
 
-    if($ResultSet){
-        print("REGISTRO ELIMINADO");
-    } else {
-        print("REGISTRO NO ELIMINADO");
-    }
-?>
+                if (mysqli_num_rows($checkResult) > 0) {
+                
+                    $SQL = "DELETE FROM CENTROSVERIFICACION WHERE ID = '$IdPropietario'";
+                    $ResultSet = Ejecutar($Con, $SQL);
+
+                    if ($ResultSet) {
+                        echo "<p style='color: green;'>REGISTRO ELIMINADO</p>";
+                    } else {
+                        echo "<p style='color: red;'>ERROR AL ELIMINAR EL REGISTRO</p>";
+                    }
+                } else {
+                  
+                    echo "<p style='color: red;'>NO SE ENCONTRÓ EL REGISTRO CON EL ID PROPORCIONADO</p>";
+                }
+
+                Desconectar($Con);
+            } catch (mysqli_sql_exception $e) {
+                echo "<p style='color: red;'>ERROR: No se puede eliminar el registro debido a restricciones de clave foránea.</p>";
+            }
+        }
+        ?>
