+++ php forms/delete/DVehiculos.php
@@ -1,17 +1,29 @@
 <?php
-    //Obtener datos
+if ($_SERVER['REQUEST_METHOD'] == 'POST' && isset($_POST['NUMSERIE'])) {
     $Id = $_POST['NUMSERIE'];
 
-    $SQL= "DELETE FROM VEHICULOS WHERE NUMSERIE = '$Id'";
-
     include("../../controlador/controlador.php");
-    $Con = Conectar();
-    $ResultSet = Ejecutar($Con,$SQL);
-    Desconectar($Con);
 
-    if($ResultSet){
-        print("REGISTRO ELIMINADO");
+    $Con = Conectar();
+
+    $checkSQL = "SELECT * FROM VEHICULOS WHERE NUMSERIE = '$Id'";
+    $checkResult = Ejecutar($Con, $checkSQL);
+
+    if (mysqli_num_rows($checkResult) > 0) {
+        $SQL = "DELETE FROM VEHICULOS WHERE NUMSERIE = '$Id'";
+        $ResultSet = Ejecutar($Con, $SQL);
+
+        if ($ResultSet) {
+            echo "REGISTRO ELIMINADO";
+        } else {
+            echo "ERROR AL ELIMINAR EL REGISTRO";
+        }
     } else {
-        print("REGISTRO NO ELIMINADO");
+        echo "NO SE ENCONTRÓ EL REGISTRO CON EL NUMERO DE SERIE PROPORCIONADO";
     }
+
+    Desconectar($Con);
+} else {
+    echo "ERROR: No se proporcionó el número de serie.";
+}
 ?>
